+++ src/main/java/pl/simpleascoding/tutoringplatform/service/jwt/JwtServiceImpl.java
@@ -5,6 +5,7 @@ import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -12,7 +13,8 @@ import java.util.Date;
 import static pl.simpleascoding.tutoringplatform.common.SecurityFinals.*;
 
 @Service
-public class JwtServiceImpl implements JwtService {
+@Primary
+class JwtServiceImpl implements JwtService {
     private final Algorithm algorithm;
 
     @Value("${jwt.accessLifetime}")
