+++ src/main/java/pl/simpleascoding/tutoringplatform/service/review/ReviewModelMapper.java
@@ -1,11 +1,11 @@
-package pl.simpleascoding.tutoringplatform.mapper;
+package pl.simpleascoding.tutoringplatform.service.review;
 
 import org.mapstruct.Mapper;
 import pl.simpleascoding.tutoringplatform.domain.review.Review;
 import pl.simpleascoding.tutoringplatform.dto.ReviewDTO;
 
 @Mapper(componentModel = "spring")
-public interface ReviewModelMapper {
+interface ReviewModelMapper {
 
     ReviewDTO mapReviewToDto(Review review);
